Separating Container Components From Presentational Components
