+++ 07week/react-tictactoe/src/App.test.js
@@ -4,6 +4,6 @@ import App from './App';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<TicTacToeContainer />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
