+++ 07week/react-tictactoe/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import TicTacToeContainer from './Components/ContainerComponent';
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(<TicTacToeContainer />, document.getElementById('root'));
 registerServiceWorker();
