+++ 07week/react-tictactoe/src/styles.js
@@ -1,0 +1,55 @@
+const rowStyle = {
+  height: '110',
+  fontSize: '100px',
+  margin: '15px auto',
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'center'
+}
+
+const boxStyle = {
+  height: '110px',
+  width: '110px',
+  backgroundColor: 'lightblue',
+  margin: 'auto 20px',
+  border: '5px solid black',
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'center'
+}
+
+const buttonStyle ={
+  height: '50px',
+  width: '100px',
+  backgroundColor: 'grey',
+  border: '2.5px solid black',
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'center'
+}
+
+const headerStyle = {
+  fontSize: '75px',
+  backgroundColor: "grey",
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'center'
+}
+const footerStyle = {
+  height:'100px',
+  fontSize: '50px',
+  backgroundColor: "grey",
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'center'
+}
+const background = 'pink url("http://www.clker.com/cliparts/6/8/Z/b/2/1/tic-tac-toe-md.png") fixed';
+
+export const styles = {
+  rowStyle: rowStyle,
+  boxStyle: boxStyle,
+  buttonStyle: buttonStyle,
+  headerStyle: headerStyle,
+  footerStyle: footerStyle,
+  urlStyle: background
+}
